--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/BackIdle_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 283729ee7c118d14595f03265d41b0c9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
