--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/BackSit_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3192ba9590f6dc64aad5c249e5639724
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
