--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/BackWalk_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7caf56779f684974188372cbf0d3974b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
