+++ mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/BlackRabbit.controller.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 0ee7d9dacf70d884b99778c5b12f0927
-DefaultImporter:
+guid: abe490d8b14d53d4c9b0751b7e5080c9
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
