--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/FrontIdle_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 275c76ea6f461fb4bbc1cd865dcf407f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
