--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/FrontSit_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ee0db488866c60942b89b8a9edf8125e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
