--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/FrontWalk_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5b01b6875b63b3242981fd0ace6b717a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
