--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/RightIdle_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ea0ebfc7199865e4c86b88d37c206bb5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
