--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/RightSit_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9fa45ffe354cc144dae1b904d2e40861
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
