--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/Black/Animations/RightWalk_Black.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c20bdff5f2918e344a5c2f54dae865ee
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
