--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/BackSit_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 779363620614f0e4395f263708939e16
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
