--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/BacktWalk_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dcc5303a49b560e409d990763dfad4aa
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
