--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/FrontSit_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e2bd23d827372f499636e35773be950
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
