--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/FrontWalk_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a23bf17374761e041980d347d119334e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
