--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/RightIdle_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 34ab7be0923443d4ca7431a3c8f80c14
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
