--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/RightSit_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1c6f4da2f0f9ef948a2b22aff0fea00f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
