--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/RightWalk_White.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a4afe22c50151a64baf1eba573653cf7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
