--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/Walk.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d42b610527c33c3438131c64fd57f3a6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
