--- mylittlekitten/Assets/Cafe/Cafe1.1/Rabbits/White/Animations/WhiteRabbit_BackIdle.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4fa1c14f811d28d43bad648bb2de6bb8
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
