+++ mylittlekitten/Assets/Prefabs/WhiteRabbit.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 9a16442ce6c29f0489a85ad4f2ff2a53
-NativeFormatImporter:
+guid: 1879e8dab110ef0499ad3727fd58075c
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
